Real shopping

+++ locustfile.py
@@ -1,4 +1,5 @@
 import base64
+import random
 
 from locust import HttpUser, TaskSet, task
 from random import randint, choice
@@ -8,6 +9,8 @@ class WebTasks(TaskSet):
 
     @task
     def load(self):
+        tx_id = random.randint(1, 1000000)
+        print(f"[start]\ttx_id={tx_id}")
         x = bytes('%s:%s' % ('user', 'password'), 'utf-8')
         base64string = base64.b64encode(x).decode()
 
@@ -19,10 +22,21 @@ class WebTasks(TaskSet):
         self.client.get("/login", headers={"Authorization":"Basic %s" % base64string})
         self.client.get("/category.html")
         self.client.get("/detail.html?id={}".format(item_id))
-        self.client.delete("/cart")
-        self.client.post("/cart", json={"id": item_id, "quantity": 1})
-        self.client.get("/basket.html")
-        self.client.post("/orders")
+
+        # Add cart: 11618 / 442219
+        x = random.randint(1, 100)
+        if x <= 2.62:
+            self.client.delete("/cart")
+            self.client.post("/cart", json={"id": item_id, "quantity": 1})
+            self.client.get("/basket.html")
+        else:
+            return
+
+        # Order: 4621 / 11618
+        x = random.randint(1, 100)
+        if x <= 39.77:
+            print(f"[purchace]\t{tx_id}")
+            self.client.post("/orders")
 
 
 class Web(HttpUser):
